renames ListingTypes controller to AllListingTypes and added ActiveListingTypes.
Need to add a required session key parameter to all existing and future
controlers with the exception of the auth controller that will return
the session key

Session keys will renew with each api call but will expire after x
minutes of idle. All of that will be added later.